Make the 64-bit version work in JavaScript
To work around the fact that JavaScript doesn't support i64, the seed is
stored in the first word (64-bits) of the given memory and the actual
data follows. The result is then stored where the seed was stored, so
the 64-bit hash can be created in JavaScript by concatenating the hex
values of the two u32 bits.