Change API to a function returning a promise
It makes more sense that the WebAssembly instance is the promise the
user will await and then the hash functions themselves are synchronous.
This avoids creating unnecessary promises and it turns out to be faster.

Before, a class was used, because I wanted to include a scheduler as the
hash functions (h32 and h64) returned a promise so it would have been
possible that they overwrite the memory that WebAssembly is using.