Convert the package to type: module with CJS fallback
Use ES Modules by default and opt out of it by using the explicit `.cjs`
extension for tools that don't support ES modules.

That also means that the published package will automatically select the
cjs or esm version depending on which kind of module system the user is
using. It is defined in the `package.json` field `exports`.