Replace bundlesize with bunldemon and add Node 17 to CI
Node 17 is the latest release and should be tracked in CI, but
bundlesize could not be installed because it had a dependency of
`iltorb`, which cannot be compiled on Node 17. Both of these are
outdated.
There are other bundle size watchers, specifically `bundlewatch`, which
seem to be more polished, but that seems to just send a report to one of
their servers, which I don't want, so I opted for something that
requires no additional configuration or fiddling to get a very simple
output. `bundlemon` would also support GitHub Actions, but it was never
the point to be so excessive about the bundle size, it's more to just
see it, if necessary/being curious.